[GC] Fix collection_weight metrics
Integer division yield 0-1 values. This fixes by converting to double.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5731692
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95187}

+++ src/heap/gc-tracer.cc
@@ -1598,10 +1598,11 @@ void GCTracer::ReportFullCycleToRecorder() {
       event.main_thread_collection_weight_cpp_in_percent = 0;
     } else {
       event.collection_weight_cpp_in_percent =
-          event.total_cpp.total_wall_clock_duration_in_us /
+          static_cast<double>(event.total_cpp.total_wall_clock_duration_in_us) /
           total_duration_since_last_mark_compact_.InMicroseconds();
       event.main_thread_collection_weight_cpp_in_percent =
-          event.main_thread_cpp.total_wall_clock_duration_in_us /
+          static_cast<double>(
+              event.main_thread_cpp.total_wall_clock_duration_in_us) /
           total_duration_since_last_mark_compact_.InMicroseconds();
     }
   }
@@ -1727,10 +1728,10 @@ void GCTracer::ReportFullCycleToRecorder() {
     event.main_thread_collection_weight_in_percent = 0;
   } else {
     event.collection_weight_in_percent =
-        event.total.total_wall_clock_duration_in_us /
+        static_cast<double>(event.total.total_wall_clock_duration_in_us) /
         total_duration_since_last_mark_compact_.InMicroseconds();
     event.main_thread_collection_weight_in_percent =
-        event.main_thread.total_wall_clock_duration_in_us /
+        static_cast<double>(event.main_thread.total_wall_clock_duration_in_us) /
         total_duration_since_last_mark_compact_.InMicroseconds();
   }
